Create cycle_3____t_test_two_tailed_test.py
(Two-Tailed t-Test):
A nutritionist wants to determine whether a new diet plan has an effect (either increase or decrease) on the average weight of participants. The average population weight before the diet was 20 kg. A sample of 30 individuals who followed the new diet plan has a mean weight of 25 kg with a standard deviation of 4. At a 5% significance level, test whether the new diet has caused a significant change in average weight